apps/desktop/backend/windows/requirements.txt
bleak
llama-cpp-python
winrt-runtime
winrt-Windows.Devices.Bluetooth
winrt-Windows.Storage.Streams
pyinstaller
bleak